SwitchConfig.pm: allow bonding interfaces to be hardware interfaces
On platforms that opt-in to the support for hardware bonding, consider
bonding interfaces to be hardware interfaces. At the moment this is a
static per-platform decision, but in the future could be enhanced to
be more dynamic in the future should the use case appear.

Consider the same set of features as requiring platform support on
dataplane interfaces as requiring platform support on hardware bonding
interfaces. This allows the platform to opt-in to support for some,
but not necessarilly all of the functionality capable of being
configured on the interface.

VRVDR-49261