Add interfaces-switch yang deviations on siad platform
The mac address of a switch can not be supported properly on a
siad platform due to limitations of the hardware. Only the 12
least significant bits of the mac address can be modifed, and
this means that most values will not be supported.

Mark the command as deprecated on the siad platforms. Ignore the
value if it is configured.

This restriction is not added for the broadcom flexware platform,
even though they have the same HW restriction, as they do L3
forwarding in software.

VRVDR-49328