Link flap when adding IPv6 address
The dataplane is notified of interface speed whenever there is a
change to the interface configuration, even if this unrelated to
speed or duplex setting. The reason is that the handling for the
interdependent speed and duplex settings is in vyatta-intf-end, which
is called for any interface changes. The solution is to only modify
cstore if these settings have changed. Neither the cstore nor the
dataplane are the source of this config and cannot be queried in any
case for the old settings. A long-term feature VRVDR-38795 is to
improve interface mgmt, but a solution is needed before then.

VRVDR-49343