vyatta-switch: set hw-switching enabled for router platforms
The current set_default_state implementation depends on transitions to
and from being members of a bridge interface in order to trigger
changes, including the setting of hardware-switching enabled, and on
switch platforms this is also what sets the initial hardware-switching
cstore state for an interface (in the absence of any explicit
configuration). However, on router platforms the initial state is that
router interfaces are not a member of a bridge/switch so there is no
initial transition that triggers the hardware-switching
enabled/disabled cstore to be set, and this results in incorrect state
in the dataplane for functionality that depends on this.

Therefore, set the hardware-switching cstore state on a router
platform, even when the interface isn't already a member of a
bridge/switch. In order to prevent redundant updates every time some
interface config is changed (which cause extra work throughout the
system) then store a cookie value to note that the initial setting has
been done which is then used when outside of the to/from bridgr/switch
member transitions to not issue further hardware-switching cstore
updates.

Also, now that we have hardware platforms that default to router type
avoid the assumption that router platform type means
hardware-switching should be disabled, but instead pass the
platform-default platform value down and assume hardware-switching
should default to enabled when the platform type is router and this is
also the default platform type.

Fixes: VRVDR-49515