vyatta-switch: always set hw-switching state for hardware interfaces
An interface is determined as not being defaulted if it has bond-group
or address configuration on it, since it shouldn't be implicitly added
to a switch on a switch platform. However, the setting of hw-switching
state is also omitted in this case and the dataplane relies on this
signal to determine whether or not to program state down to the FAL,
with the default being hardware switching disabled.

Therefore, when setting default state enumerate over all hardware
interfaces and if one isn't defaulted then just set hardware switching
state, which then gives the system the required positive indication of
whether or not hardware-switching is enabled or not for the interface.

Fixes: VRVDR-49830