GRE: Tunnel parameters setting should be deferred
Given the deferral of tunnel creation and more importantly re-creation
of tunnels in cases where the encapsulation changes, any tunnel
parameters that need to be set based on the tunnel type also need to be
deferred.  This is so that the correct tunnel type can be used.  If
this is not done, then the config_create script uses the old tunnel
encapsulation to determine the type, i.e. prior to the cli change,
and causes script failures.

VRVDR-47858