lib: fix getIP error message
in some cases ip addr may return success but with a warning
'Dump was interrupted and may be inconsistent.'. /bin/ip exit
with 0 on Dump error.
This is usually not a major issue and can happen when there are
changes happening to the interfaces while the ip commands initial
netlink dump was going on.

This can cause gettree /interface/statistics to fail as configd
can't handle this error message and logs the following:

/usr/sbin/configd[2706]: Warnings generated creating full config
and state tree:
/usr/sbin/configd[2706]: Error: /interfaces/statistics: Invalidly
formatted data returned for (*schema.container)statistics: Dump
was interrupted and may be inconsistent.

Change the code to suppress the error message on success.

VRVDR-50454

(cherry picked from commit c83451f9cd42fbd86ea86d6d1cdc8bcbbdb317b1)