Update VRRP RFC mapping to use VRRP VCI RPC
Moving vyatta-vrrp to use the new VCI component also brought in an RPC
that can be used to query the base interface and group ID for a VRRP RFC
interface.
Updated Interface.pm to use this new RPC.
When keepalived has shutdown we can get an error on the commit that
removes all the VRRP config, added a special case where if the RPC
returns a vrid of 0 just return as there's no VRRP config.

Closes:VRVDR-51492