SwitchConfig.pm: write ManagementPorts value to switch.conf
It may be useful for the rest of the system to know on a given
platform which interfaces are intended for use as management ports,
and to be able to know what name was picked for the interface in the
system, rather than just the PCI address.

Thus use the ifconfig results from the dataplane along with a new
field exposed by the dataplane to build a list of management ports and
write this to switch.conf.

Closes: VRVDR-52175