switch: ensure disable_ipv6 is deleted when switch-port is deleted
When an interface is unlinked from a switch and, at the same time, the
switch is deleted, the "disable_ipv6" attributes remains attached to
the interface.

Since the switch has a higher (YANG) priority it is removed from the
system before the interface switch-group is processed leaving the
"disable_ipv6" sysctl untouched.

Adjust the core update_interfaces() function to ensure that when an
interface is not linked to a switch, the "disable_ipv6" attribute is
updated (deleted) as necessary.

Fixes: VRVDR-45948