Platform.pm: get current platform type from candidate tree
The validation of functionality configured on interfaces is currently
mostly inverted when changing platform type since the validation uses
a value stored in a configuration file and the configuration file is
only updated during commit time.

Fix this by changing get_platform_type_and_default which is used as
part of feature platform validation to use the candidate configuration
for "system platform type" instead.

Fixes: VRVDR-52666