vyatta-switchport: fix interfaces spuriously added to switch on router
When a platform is configured to be a router and yet there is still a
switch configured with a physical-switch value, then interfaces
without switch configuration adding them to that switch should be left
as standalone interfaces, but that isn't the case currently.

Fix interfaces without explicit switch configuration being added as
members of the switch when the physical-switch configuration is
processed by bailing early from update_physical_switch.

Fixes: VRVDR-53052