vyatta-show-interfaces: add an option to show platform state for vifs
We want to be able to show the dataplane platform state for vifs so
add an option into the script to do that. This will call out to other
scripts to query the state and print it.

The dataplane show command for a vif should ideally call a script in
the same repo as the yang is in, but instead it calls
vyatta-show-interfaces.pl. To be able to add the 'platform state' to
the end of the output there are several options:
1) Change the current output for a vif to make it more inline with
the top level interface output. This is bad as users may be relying on
the current output.

2) Take a copy of the show code here, and put it in the same repo as
the top level yang that shows a dataplane vif is in. This is also bad
as then we will have mutliple copies of the code that will get out of
sync.

3) Add a new option to the existing script that will print out the
current output but then call out to other scripts to print the
'platform state' part of the output.  This is the preferred option.

VRVDR-46947