Interface.pm: return a valid from the vlan sub when vlan not configured
In the case that a vlan isn't explicitly configured for a vif then the
vlan will default to that of the vif number. However, the
Interface::vlan sub returns an undefined value in this case.

So make it more friendly to the callers of this function by returning
a valid vlan of that of the vif in this case.

VRVDR-53560