Bugfix/Check for key existence in RPC return
Infrequently a VRRP RPC can return an empty dictionary if it is called
before the interface exists, this causes warnings to be printed on the
CLI. There's no material effect on configuration but the problem should
still be dealt with.
Changed comparison operator from eq to ==, the returned RPC value is an
int not a string.

Closes:VRVDR-54767