Improve efficiency of VIF validation.
Even using must statements instead of scripts, validation of VIFs on an interface with many VIFs is slow. On a vRouter software-only ISO, 300 VIFs on a bonding interface takes almost 30s, of which over 20s is running must statements.  The time is O(n squared) as each VIF's must statement gathers information on every other VIF.

The fix is to use custom XPATH functions, and to move the validation from per-VIF to per-interface when the custom function is available via configd:must on-box. Off-box validation will continue to use the existing must statements.

Essentially we move from:

list bonding {
	 list vif {
	 	  must "expensive-operation-run-for-each-vif()";
}

to this:

list bonding {
	 must "true()" {
	 	  configd:must "cheaper-operation-run-once-per-interface()";
	 }
     list vif {
          must "expensive-operation-run-for-each-vif()" {
		  	   configd:must "true()";
		 }
}

While there is a risk that a new interface type will use the VIF grouping without adding the validate-vif-vlan-settings() call on the interface, it's rare for a new interface type to be created, and there are comments on the VIF groupings, and anyone copying an existing interface using VIFs will see the configd:must anyway.  Note that this change is to vif-group (used for bonding, dataplane and switch). vhost interfaces use vif-group-body which is augmented, so we can't put a must statement there at the interface level (illegal to have must directly under an augment). So for this grouping we use more efficient custom functions, but have to run them for each VIF.

As an example of the improvements seen, validation of 300 VIFs on a bonding interface on a VNF ISO reduces from 26s to 4s, with only 250ms now due to must statements (previously > 20s).

I have an existing set of interface-validation acceptance tests that verify pass and fail cases for these must statements, including bonding, switch, vhost and dataplane interfaces.

Separate PRs will be put out for dataplane, bonding and vhost interfaces, and a single DM review will be created.

NB: while we lose some information on the output error (no longer know which VIF is the problem) we avoid duplicated information where we get effectively the same error for both / all VIFs that have the duplicated vlan id(s), and the time saving is essential for the device to be responsive. A 30s+ hit on every validation simply isn't acceptable.

Part of VRVDR-53594.

(cherry picked from commit 276fa70a93c37540b9cee1f565395443d711aecf)