Fix issue of IPv6 tunnel being brought down/up when base config unchanged
When adding or removing features to a tunnel which is using IPv6 as a
local address, the tunnel is tore down and created again.  This also
means it gets assigned a new ifindex value.

This occurs due to a bug in the vyatta-tunnel-deferred.pl script.
For IPv6, if the new encap is "gre" it changes it to "ip6gre". However,
it does not change the old encap in a similar manner. Later on it checks
to see if the encap has changed by comparing the old one ("gre") with
the new one ("ip6gre"). As they are different it thinks the encap has
changed, and so reconfigures the tunnel causing the tunnel interfaces
to be brought down and up again.

VRVDR-56190